Added IMAT config
